--- packages/kokoas-server/src/handleRequest/reqSendContractDirectV2/getContractFlow/getContractFlow.ts
@@ -1,43 +1,0 @@
-import { ReqSendContractParams } from 'types';
-import { TContractData } from '../getContractDataV2';
-
-export const getContractFlow = ({
-  data,
-}:{
-  data: TContractData,
-}) => {
-
-  const {
-    customers,
-    cocoAG,
-    projName,
-
-    storeMngrName,
-    storeMngrEmail,
-
-    accountingName,
-    accountingEmail,
-
-    mainAccountingName,
-    mainAccountingEmail,
-
-    subAccountingName,
-    subAccountingEmail,
-
-    projTypeName,
-    
-    signMethod,
-    
-  } = data;
-
-  // electronic
-
-  // wetInk
-
-  // jisha
-
-  // sabisu
-
-
-
-};
